Vue component: BeaconController
